+++ CHANGELOG.md
@@ -8,12 +8,14 @@ Most recent version numbers *should* follow the [Semantic Versioning](https://se
 
 ### Added
 - double angle mapping (DAM) B1 mapping option
+- option to test ambiguous angle ranges beyond [0,90] and [90,180] degrees for SE/STE B1 mapping (nAmbiguousAngles b1 defaults parameter)
 - various B1 input checking
 
 ### Fixed
 - issue #42 (Non-matching filename length can cause AFI B1 calculation crash)
 - issue #46 (Toolbox expects SE/STE input sorted by echo and flip angle whereas BIDS sorts by flip angle then echo)
 - issue #26 (Possible bug in calculation of B1 maps with SE/STE EPI)
+- SE/STE B1 mapping calculation can now no longer select both ambiguous angles for a given SE/STE pair
 
 ## [v0.4.0]